code optimize

+++ linkis-public-enhancements/linkis-pes-publicservice/src/main/java/org/apache/linkis/basedatamanager/server/restful/ErrorCodeRestfulApi.java
@@ -89,8 +89,12 @@ public class ErrorCodeRestfulApi {
   @ApiOperation(value = "remove", notes = "Remove an Error Code by id", httpMethod = "DELETE")
   @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
   public Message remove(HttpServletRequest request, @PathVariable("id") Long id) {
-    ModuleUserUtils.getOperationUser(
-        request, "Remove a Datasource Code Record,id:" + id.toString());
+    String username =
+        ModuleUserUtils.getOperationUser(
+            request, "Try to remove error code record with id:" + id.toString());
+    if (!Configuration.isAdmin(username)) {
+      return Message.error("User '" + username + "' is not admin user[非管理员用户]");
+    }
     boolean result = errorCodeService.removeById(id);
     return Message.ok("").data("result", result);
   }
